% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildFunctions.R
\name{bubbleHeatmapList}
\alias{bubbleHeatmapList}
\title{Output List of bubbleHeatmap Plots.}
\usage{
bubbleHeatmapList(inputList)
}
\arguments{
\item{inputList}{A list of two lists, trees and legend. trees should contain
a list of lists of arguments for multiple calls to bubbleHeatmap and legend
should contain a list of arguments for a call to bubbleHMLegends. This
object can be generated automatically using \code{\link{multiPlotInput}}.}
}
\value{
A list of bubbleHeatmap \code{\link[grid]{gTree}} objects. Any
legends specified by the legend list will also be produced as separate trees
in the output list.
}
\description{
Convenient wrapper for \code{bubbleHeatmap} and \code{bubbleHMLegend} for
easy output of multiple plots from lists of arguments.
}
\seealso{
Other build functions: 
\code{\link{bubbleHMLegends}()},
\code{\link{bubbleHeatmap}()},
\code{\link{multiPlotInput}()}
}
\concept{build functions}
