% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrangeFunctions.R
\name{getLayoutSizes}
\alias{getLayoutSizes}
\title{Recalculate Layout Sizes for bubbleHeatmap gTree}
\usage{
getLayoutSizes(tree, unitBase)
}
\arguments{
\item{tree}{A bubbleHeatmap \code{\link[grid]{gTree}}.}

\item{unitBase}{The unitBase parameter used when the tree was created.}
}
\value{
The input bubbleHeatmap \code{\link[grid]{gTree}} with viewport
dimensions matched to the size of the children grobs.
}
\description{
This function recalculates the viewport layout dimensions for a bubbleHeatmap
tree based on the sizes of the children grobs, and returns the updated tree.
This is useful for matching widths exactly to grob dimensions in an updated
drawing context, and for resetting any previous layout modifications.
}
