% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatData.R
\name{merge_template}
\alias{merge_template}
\title{Merge Nightingale Figure Templates with User Data}
\usage{
merge_template(data, id)
}
\arguments{
\item{data}{User dataset}

\item{id}{Name of key column, either "ukbb_id" or "ckb_id"}
}
\value{
Merged dataset
}
\description{
Custom merge function to help users align their data with the template
datasets containing the variable location information for the Nightingale
figures. User datasets should contain the variables representing bubble color
and size, and a key column to merge rows with the template. The key column
should either be named "ukbb_id" and contain the UK Biobank Field ID for each
variable (e.g. "23474") or named "ckb_id" and contain the China Kadoorie
Biobank variable ID, (e.g. "nmr_bohbut").
}
\examples{
data <- merge_template(cetp, "ckb_id")
}
