% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nightingale.R
\name{nightingalePlot}
\alias{nightingalePlot}
\title{Arrange List of bubbleHeatmap Plots According to Preset Nightingale Template.}
\usage{
nightingalePlot(treeList)
}
\arguments{
\item{treeList}{List of bubbleHeatmap trees. This function expects a treeList
of particular length, containing trees with specified names and dimensions.
The easiest way to generate this is via the \code{\link{formatData}},
\code{\link{multiPlotInput}} and \code{\link{bubbleHeatmapList}} functions.
See vignette for full details of this workflow.}
}
\value{
A \code{\link[grid]{gTree}} object representing the input
bubblehHeatmaps arranged into a Nightingale metabolomics figure.
}
\description{
Produce a single \code{\link[grid]{gTree}} object representing plots arranged
according to the provided Nightingale metabolomics figure template.
}
